refactor: streamline comment handling and enhance PostCard components
- Removed the CommentsModal component and integrated PostCommentsSection for better modularity.
- Updated PostCardFooter to manage comment visibility and interactions more effectively.
- Enhanced PostModal to support scrolling to comments upon opening.
- Improved state management for comments across PostCard and PostCardMainContent components.
